skills/requirements.txt
# LexFlow Skills - Dependências
# ============================================================

# Core
pyyaml>=6.0.1
python-dotenv>=1.0.0

# LexFlow SDK (se disponível publicamente)
# lexflow-sdk>=1.0.0

# Utilitários
click>=8.1.0          # CLI framework
rich>=13.0.0          # Terminal formatting
httpx>=0.25.0         # HTTP client async

# Desenvolvimento
pytest>=7.4.0         # Testing
black>=23.0.0         # Code formatting
ruff>=0.1.0           # Linting

# Opcional - Para funcionalidades avançadas
jinja2>=3.1.0         # Template engine
pydantic>=2.0.0       # Data validation

# GitHub integration (para PRs de opcodes)
PyGithub>=2.1.0       # GitHub API client